Use ProjectBuilder instead of dryRun for DetektPlainSpec
